requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==0.21.1
gunicorn==22.0.0
celery==5.4.0
redis==5.0.6
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
firebase-admin==6.5.0
pymysql==1.1.1
